use long long int in execution time table

+++ include/communication.hpp
@@ -11,22 +11,22 @@ public:
     Communication() {}
     virtual ~Communication() {}
 
-    virtual void GetCommunicationTable(std::shared_ptr<DAG>& dag, int numberOfRunnables, int hyperPeriod, std::vector<std::vector<std::vector<int>>>& runnablePermutation, std::vector<std::vector<std::vector<ExecutionInformation>>>& runnableCommunications) = 0;
+    virtual void GetCommunicationTable(std::shared_ptr<DAG>& dag, int numberOfRunnables, long long int hyperPeriod, std::vector<std::vector<std::vector<int>>>& runnablePermutation, std::vector<std::vector<std::vector<ExecutionInformation>>>& runnableCommunications) = 0;
 };
 
 class RunnableImplicit : public Communication {
 public:
-    void GetCommunicationTable(std::shared_ptr<DAG>& dag, int numberOfRunnables, int hyperPeriod, std::vector<std::vector<std::vector<int>>>& runnablePermutation, std::vector<std::vector<std::vector<ExecutionInformation>>>& runnableCommunications);
+    void GetCommunicationTable(std::shared_ptr<DAG>& dag, int numberOfRunnables, long long int hyperPeriod, std::vector<std::vector<std::vector<int>>>& runnablePermutation, std::vector<std::vector<std::vector<ExecutionInformation>>>& runnableCommunications);
 };
 
 class TaskImplicit : public Communication {
 public:
-    void GetCommunicationTable(std::shared_ptr<DAG>& dag, int numberOfRunnables, int hyperPeriod, std::vector<std::vector<std::vector<int>>>& runnablePermutation, std::vector<std::vector<std::vector<ExecutionInformation>>>& runnableCommunications);
+    void GetCommunicationTable(std::shared_ptr<DAG>& dag, int numberOfRunnables, long long int hyperPeriod, std::vector<std::vector<std::vector<int>>>& runnablePermutation, std::vector<std::vector<std::vector<ExecutionInformation>>>& runnableCommunications);
 };
 
 class LET : public Communication {
 public:
-    void GetCommunicationTable(std::shared_ptr<DAG>& dag, int numberOfRunnables, int hyperPeriod, std::vector<std::vector<std::vector<int>>>& runnablePermutation, std::vector<std::vector<std::vector<ExecutionInformation>>>& runnableCommunications);
+    void GetCommunicationTable(std::shared_ptr<DAG>& dag, int numberOfRunnables, long long int hyperPeriod, std::vector<std::vector<std::vector<int>>>& runnablePermutation, std::vector<std::vector<std::vector<ExecutionInformation>>>& runnableCommunications);
 };
 
 #endif
